simple-rag/requirements.txt
anthropic
sentence-transformers
numpy
python-dotenv
pymupdf
openai